requirements-dev.txt
-r requirements.txt

astroid==2.12.13
colorama==0.4.6
dill==0.3.6
dodgy==0.2.1
flake8==5.0.4
flake8-polyfill==1.0.2
gitdb==4.0.10
GitPython==3.1.30
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
packaging==21.3
pep8-naming==0.10.0
platformdirs==2.6.2
poetry-semver==0.1.0
prospector==1.8.3
pycodestyle==2.9.1
pydocstyle==6.2.3
pyflakes==2.5.0
pylint==2.15.9
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pyparsing==3.0.9
PyYAML==6.0
requirements-detector==1.0.3
setoptconf-tmp==0.3.1
smmap==5.0.0
snowballstemmer==2.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
wrapt==1.14.1
factory_boy==3.2.1
coverage==7.0.1
pre-commit==2.21.0